Moving artifact repository password encryption logic to clients
